backend/requirements.txt
flask
flask-SQLAlchemy
flask-mail
requests
beautifulsoup4
python-dotenv
cors
flask-cors
langchain_groq
langchain-community
langchain_experimental
langchain-ai21
chromadb
sentence-transformers
huggingface_hub 
transformers
requests
beautifulsoup4
crawl4ai
psycopg2